Server/requirements.txt
fastapi
websockets
openai
hume == 0.6.0
numpy
ruff
python-dotenv
